+++ packages/ui/useScroll/useScroll.js
@@ -149,6 +149,8 @@ const useScrollBase = (props) => {
 
 	const [isHorizontalScrollbarVisible, setIsHorizontalScrollbarVisible] = useState(horizontalScrollbar === 'visible');
 	const [isVerticalScrollbarVisible, setIsVerticalScrollbarVisible] = useState(verticalScrollbar === 'visible');
+	const [isScrollVertically, setIsScrollVertically] = useState(false);
+	const [isScrollHorizontally, setIsScrollHorizontally] = useState(false);
 
 	const mutableRef = useRef({
 		overscrollEnabled: !!(props.applyOverscrollEffect),
@@ -345,6 +347,13 @@ const useScrollBase = (props) => {
 			removeEventListeners();
 		};
 	});
+
+	useEffect(() => {
+		const bounds = getScrollBounds();
+		setIsScrollVertically(canScrollVertically(bounds));
+		setIsScrollHorizontally(canScrollHorizontally(bounds));
+	}, []); // eslint-disable-line react-hooks/exhaustive-deps
+
 
 	// scrollMode 'translate' [[
 	// TODO: consider replacing forceUpdate() by storing bounds in state rather than a non-
@@ -1565,7 +1574,9 @@ const useScrollBase = (props) => {
 	return {
 		scrollContentWrapper: noScrollByDrag ? 'div' : TouchableDiv,
 		isHorizontalScrollbarVisible,
-		isVerticalScrollbarVisible
+		isVerticalScrollbarVisible,
+		isScrollVertically,
+		isScrollHorizontally
 	};
 };
 
@@ -1615,7 +1626,9 @@ const useScroll = (props) => {
 	const {
 		scrollContentWrapper,
 		isHorizontalScrollbarVisible,
-		isVerticalScrollbarVisible
+		isVerticalScrollbarVisible,
+		isScrollVertically,
+		isScrollHorizontally
 	} = useScrollBase({
 		...props,
 		assignProperties,
@@ -1641,7 +1654,9 @@ const useScroll = (props) => {
 		scrollContentWrapper,
 		scrollContentHandle,
 		isHorizontalScrollbarVisible,
-		isVerticalScrollbarVisible
+		isVerticalScrollbarVisible,
+		isScrollVertically,
+		isScrollHorizontally
 	};
 };