+++ packages/ui/CHANGELOG.md
@@ -43,6 +43,7 @@ No significant changes.
 ### Changed
 
 - `ui/Transition` prop `duration` to support any valid CSS value for `slide` and `fade` `type`
+- `ui/useScroll` to manage internal state `isScrollVertically` and `isScrollHorizontally`
 
 ## [3.4.11] - 2020-12-11